Extracting callback to separate controller, callback can be hit multiple times.
Removing the need for the skip_before_filters by moving the callback
action to its own controller.

Callback can be hit multiple times, and the request will succeed. DPS
always pings the callback in addition to the user being redirected to
the callback, so the callback action must be forgiving of being hit
twice.